pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "joycaption_beta_one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate",
    "omegaconf>=2.3.0",
    "peft>=0.16.0",
    "pillow",
    "torch",
    "torchaudio",
    "torchvision",
    "transformers",
    "triton",
    "wandb>=0.21.0",
    "click>=8.1.7",
    "notebook",
    "ipykernel",
    "pyarrow>=21.0.0",
]

[project.scripts]
train-accelerate = "joycaption_beta_one.finetuning.train_accelerate:main"
joycaption-app = "joycaption_beta_one.gradio_app.app:main"
convert-to-parquet = "joycaption_beta_one.finetuning.convert_to_parquet:main"

[project.optional-dependencies]
app = [
    "huggingface_hub >= 0.30.1",
    "sentencepiece",
    "liger-kernel >= 0.5.9",
    "gradio >= 5.24.0",
]

[tool.hatch.build]
packages = ["src/joycaption_beta_one"]
exclude = ["images/", "*.pt"]
